[CDH-40374] Backport SOLR-7374 - Backup/Restore should provide a param for specifying the directory implementation it should use
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/14421
Tested-by: Jenkins User

+++ solr/core/src/test/org/apache/solr/handler/TestRestoreCore.java
@@ -53,10 +53,11 @@ public class TestRestoreCore extends SolrJettyTestBase {
   TestReplicationHandler.SolrInstance master = null;
   SolrServer masterClient;
 
-  private static final String CONF_DIR = "solr" + File.separator + "collection1" + File.separator + "conf"
+  private static final String CONF_DIR = "solr" + File.separator + DEFAULT_TEST_CORENAME + File.separator + "conf"
       + File.separator;
 
   private static String context = "/solr";
+  private static long docsSeed; // see indexDocs()
 
   private static JettySolrRunner createJetty(TestReplicationHandler.SolrInstance instance) throws Exception {
     FileUtils.copyFile(new File(SolrTestCaseJ4.TEST_HOME(), "solr.xml"), new File(instance.getHomeDir(), "solr.xml"));
@@ -69,7 +70,8 @@ public class TestRestoreCore extends SolrJettyTestBase {
   private static SolrServer createNewSolrClient(int port) {
     try {
       // setup the client...
-      HttpSolrServer client = new HttpSolrServer(buildUrl(port, context) + "/" + DEFAULT_TEST_CORENAME);
+      //HttpSolrServer client = new HttpSolrServer(buildUrl(port, context) + "/" + DEFAULT_TEST_CORENAME);
+      HttpSolrServer client = new HttpSolrServer(buildUrl(port, context));
       client.setConnectionTimeout(15000);
       client.setSoTimeout(60000);
       client.setDefaultMaxConnectionsPerHost(100);
@@ -93,6 +95,7 @@ public class TestRestoreCore extends SolrJettyTestBase {
 
     masterJetty = createJetty(master);
     masterClient = createNewSolrClient(masterJetty.getLocalPort());
+    docsSeed = random().nextLong();
   }
 
   @Override
@@ -109,11 +112,12 @@ public class TestRestoreCore extends SolrJettyTestBase {
   @Test
   public void testSimpleRestore() throws Exception {
 
-    int nDocs = usually() ? TestReplicationHandlerBackup.indexDocs(masterClient) : 0;
+    int nDocs = usually() ? BackupRestoreUtils.indexDocs(masterClient, "collection1", docsSeed) : 0;
 
     String snapshotName;
     String location;
     String params = "";
+    String baseUrl = masterJetty.getBaseUrl().toString();
 
     //Use the default backup location or an externally provided location.
     if (random().nextBoolean()) {
@@ -129,7 +133,7 @@ public class TestRestoreCore extends SolrJettyTestBase {
 
     TestReplicationHandlerBackup.runBackupCommand(masterJetty, ReplicationHandler.CMD_BACKUP, params);
 
-    CheckBackupStatus checkBackupStatus = new CheckBackupStatus((HttpSolrServer) masterClient, null);
+    CheckBackupStatus checkBackupStatus = new CheckBackupStatus((HttpSolrServer) masterClient, DEFAULT_TEST_CORENAME, null);
     while (!checkBackupStatus.success) {
       checkBackupStatus.fetchStatus();
       Thread.sleep(1000);
@@ -140,29 +144,30 @@ public class TestRestoreCore extends SolrJettyTestBase {
     for (int attempts=0; attempts<numRestoreTests; attempts++) {
       //Modify existing index before we call restore.
 
-      //Delete a few docs
-      int numDeletes = TestUtil.nextInt(random(), 1, nDocs);
-      for(int i=0; i<numDeletes; i++) {
-        masterClient.deleteByQuery("id:" + i);
-      }
-      masterClient.commit();
-
-      //Add a few more
-      int moreAdds = TestUtil.nextInt(random(), 1, 100);
-      for (int i=0; i<moreAdds; i++) {
-        SolrInputDocument doc = new SolrInputDocument();
-        doc.addField("id", i + nDocs);
-        doc.addField("name", "name = " + (i + nDocs));
-        masterClient.add(doc);
-      }
-      //Purposely not calling commit once in a while. There can be some docs which are not committed
-      if (usually()) {
+      if (nDocs > 0) {
+        //Delete a few docs
+        int numDeletes = TestUtil.nextInt(random(), 1, nDocs);
+        for(int i=0; i<numDeletes; i++) {
+          masterClient.deleteByQuery("id:" + i);
+        }
         masterClient.commit();
+
+        //Add a few more
+        int moreAdds = TestUtil.nextInt(random(), 1, 100);
+        for (int i=0; i<moreAdds; i++) {
+          SolrInputDocument doc = new SolrInputDocument();
+          doc.addField("id", i + nDocs);
+          doc.addField("name", "name = " + (i + nDocs));
+          masterClient.add(doc);
+        }
+        //Purposely not calling commit once in a while. There can be some docs which are not committed
+        if (usually()) {
+          masterClient.commit();
+        }
       }
 
       TestReplicationHandlerBackup.runBackupCommand(masterJetty, ReplicationHandler.CMD_RESTORE, params);
-
-      while (!fetchRestoreStatus()) {
+      while (!fetchRestoreStatus(baseUrl, DEFAULT_TEST_CORENAME)) {
         Thread.sleep(1000);
       }
 
@@ -174,15 +179,16 @@ public class TestRestoreCore extends SolrJettyTestBase {
 
   @Test
   public void testFailedRestore() throws Exception {
-    int nDocs = TestReplicationHandlerBackup.indexDocs(masterClient);
+    int nDocs = BackupRestoreUtils.indexDocs(masterClient, "collection1", docsSeed);
 
     String location = createTempDir().getAbsolutePath();
     String snapshotName = TestUtil.randomSimpleString(random(), 1, 5);
     String params = "&name=" + snapshotName + "&location=" + URLEncoder.encode(location, "UTF-8");
+    String baseUrl = masterJetty.getBaseUrl().toString();
 
     TestReplicationHandlerBackup.runBackupCommand(masterJetty, ReplicationHandler.CMD_BACKUP, params);
 
-    CheckBackupStatus checkBackupStatus = new CheckBackupStatus((HttpSolrServer) masterClient, null);
+    CheckBackupStatus checkBackupStatus = new CheckBackupStatus((HttpSolrServer) masterClient, DEFAULT_TEST_CORENAME, null);
     while (!checkBackupStatus.success) {
       checkBackupStatus.fetchStatus();
       Thread.sleep(1000);
@@ -197,7 +203,7 @@ public class TestRestoreCore extends SolrJettyTestBase {
     TestReplicationHandlerBackup.runBackupCommand(masterJetty, ReplicationHandler.CMD_RESTORE, params);
 
     try {
-      while (!fetchRestoreStatus()) {
+      while (!fetchRestoreStatus(baseUrl, DEFAULT_TEST_CORENAME)) {
         Thread.sleep(1000);
       }
       fail("Should have thrown an error because restore could not have been successful");
@@ -208,7 +214,7 @@ public class TestRestoreCore extends SolrJettyTestBase {
     verifyDocs(nDocs);
 
     //make sure we can write to the index again
-    nDocs = TestReplicationHandlerBackup.indexDocs(masterClient);
+    nDocs = BackupRestoreUtils.indexDocs(masterClient, "collection1", docsSeed);
     verifyDocs(nDocs);
 
   }
@@ -222,8 +228,8 @@ public class TestRestoreCore extends SolrJettyTestBase {
     assertEquals(nDocs, response.getResults().getNumFound());
   }
 
-  private boolean fetchRestoreStatus() throws IOException {
-    String masterUrl = buildUrl(masterJetty.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME +
+  public static boolean fetchRestoreStatus (String baseUrl, String coreName) throws IOException {
+    String masterUrl = baseUrl + "/" + coreName +
         "/replication?command=" + ReplicationHandler.CMD_RESTORE_STATUS;
     final Pattern pException = Pattern.compile("<str name=\"exception\">(.*?)</str>");
